Add files via upload
This program is to display the name of the app “Document” followed by a blank line, prompt the user for the name and content for a document, save the document to the current directory, and display a message if it was successful or unsuccessful based on a specification provided. The filename of the document is the name provided by the user with “.txt” appended to it.